Add setting and getting of daily amount limit. (#1)
Add option to set the daily amount limit. We need to set the amount limit in order to perform transactions. There are a few other limits but they are not implemented. Setting amount limit is not working on prod, in order to set it on prod you need to write email to FitBank. Due to a bug, the API is not working in sandbox as well.